[lldb] Retire the SwiftASTContextReader lock manager.
This mechanism was put into place to deal with SwiftASTContexts being
torn down and replaced. While that may still happen, we no have a much
better solution for the problem of dangling type references:
CompilerType now holds a weak_ptr to the type system so it's safe to
destroy a type system.