[lldb] Thread mangling flavor together with node pointer
LLDB performs many analysis as transformations by converting a mangled
name into a demangle tree and later on back to a mangled name.
Unfortunately there's no cheap way for a demangle tree to carry its
mangled name's mangling flavor. This patch threads the mangling flavor
along with the demangle tree so when converting a demangle tree back to
a mangled name we reconstruct it with the same mangling flavor as the
original type.