Experimenting with general constexpr operators new and delete
Only supporting the current/default constant expression evaluator at
the moment.

We replace the getStdAllocator() function with getSafeAllocator() that
knows to look for any of std::allocator<T>::{allocate,deallocate} or
new, new[], delete, or delete[].

The code is also garbage as I don't really understand exactly the
evaluation model, and this is currently a quick experiment so it's
mode of design is very much "beat the code until it achieves the
desired result" rather than "do it properly".

constexpr new/delete is currently restricted to type aware allocation
because we use the type-identity parameter to identify the type being
allocated or deallocated. This is not semantically necessary, it is
simply a product of us not recording the type being allocated/deallocated
in the callframe so we have to infer it from the function being used.

This could easily be rectified if we really cared.