[Clang] Permit noescape on non-pointer types
In modern C++ we often use span, string_view or other view objects instead of
raw pointers. This PR opens the door to mark those noescape. This can be useful
to document the API contracts, for interop with memory safe languages like
Swift or Rust, and possibly in the future to implement take this into account
in codegen.

The PR also adds a feature test macro. The goal is to help avoiding warnings
when the code is compiler by earlier versions of clang that does not permit
this attribute on non-pointer types.

rdar://140196481
(cherry picked from commit 1b291ebe84df0e6ba8f7cb6a6896baefbb8fc5ed)