[clang][cas] Avoid calling freezeConfig for every cache key
With the libclang caching APIs, it is possible to configure CASOptions
separately from the CAS used for scanning. In this setup, the CAS is not
pre-cached in the CASOptions and ends up being re-opened for every cache
key, which is expensive with some plugin CAS. Instead, get the hash
identifier from the already open CAS and only freezeConfig when
canonicalizing for an actual compilation.

rdar://141555438
(cherry picked from commit 3c460cc4e43df4e43af42fbbd9e00506dc7029c5)