[BoundsSafety][NFC] Simplify the interface of `BoundsSafetyCheckAssignmentToCountAttrPtrWithIncompletePointeeTy`
Previously the interface took a std::function<std::string>. The
rationale behind this was to prevent callers from always allocating and
computing a `std::string`.

While trying to upstream this code (rdar://133600117)
(https://github.com/llvm/llvm-project/pull/106321) it was pointed out
that there might be a simpler way to implement this.

This patch instead has callers pass

* a `ValueDecl*` pointer. In the cases where this isn't known (currently
  return values and unnamed parameters) this can be set to nullptr
* a boolean flag stating whether or not the `ValueDecl*` should be fully
  qualified when printed.

This avoids needing to pass a `std::function` and avoids `std::string`
allocation.