[lldb][swift] Mark async unwinding plans as SourcedFromCompiler
LLDB always attempts to check other unwinding plans when an unwind plan
produces a CFA == 0. It does so because it doesn't distinguish the case
where the plan truly produced the end of the stack versus the case where
the plan was faulty.

We need a mechanism to disable this for swift async plans: if the
language runtime constructs such a plan, then that plan will correctly
identify the bottom of the stack. Checking other plans is harmless in
non-async settings (they would just produce the same backtrace), but it
is harmful for async virtual backtraces, as a "regular" unwind plan
would produce a real backtrace if it were asked to start from the top of
the stack (which would happen in short backtraces).

Until such mechanism exists, we can work around the issue by claiming
swift async plans are "sourced from compiler", which seem to disable the
fallback mechanism because plans sourced from the compiler are
considered the most reliable plans.

rdar://142683622