[lldb][swift] Swift OS Plugin
This commit introduces an OS plugin converting Process threads into
OperatingSystem threads that represent a Task. This is done by
inspecting ThreadLocalStorage for a specific location where the swift
runtime writes information about the current Task.

With this plugin, ThreadPlans now work properly in async code: a step
action on a Thread is now effectively an action on a Task, and LLDB will
correctly track that. ThreadPlans are effectively "Task Plans".