[Swift REPL] Inherit the environment for the Swift REPL
The Swift REPL was had rolling the ProcessLaunchInfo instead of using
the one already setup by the target. While populating the launch info,
it used `Target::GetTargetEnvironment()` which only looks at the
corresponding setting, and ignoring the `inherit-environment` setting.
The alternative of calling `Target::GetInheritedEnvironment` is also
wrong, because that has the inverse problem: it only inherits the
environment and doesn't merge it with the target's environment
variables. The solution is to use `Target::GetProcessLaunchInfo` which
populates the launch info correctly.

Propagating the environment correctly is particularly important on
Windows where `Path` needs to be passed to the inferior to allow
`LoadLibraryW(L"swiftCore.dll")` to succeed. Without this patch, the
library is not found and the inferior exits terminating the REPL
instance.

- Fixes: llvm/llvm-project#9551
- Will fix swiftlang/swift#76702 and swiftlang/swift#70005 with a
  corresponding driver change to set `inherit-environment` when invoking
  the REPL.

rdar://137522456