[lldb] Add more expression evaluation resilience tests
Add more resilience tests with a public superclass that is generic and
inherits from an objc type and a non-public subclass with a property we
access on the test. The tests vary in which modules the types exist (one
resilient executable, one resilient library and non resilient
executable, 2 resilient libraries and one resilient executable).