[-Wunsafe-buffer-usage] Add a warning gadget for single pointer argument
Add a warning gadget for passing a pointer to a __single pointer
parameter in an unsafe way. This commit adds support for recognition of
the following safe patterns:
- `nullptr`
- `&var`, if `var` is a variable identifier
- `&C[_]`, if `C` is a hardened container/view
- `span.first(1).data()` and friends
If the argument is not one of those patterns, it is considered unsafe
and a warning is emitted.

rdar://128157528

(cherry picked from commit db482302575cda64dbc65b2b297377705eaf60dc)