[lldb] Run generic expression evaluator first
The generic expression evaluator can be significantly faster than the
regular one since it only needs to import information about the module
the debugger is stopped at to evaluate the expression. This patch
swaps the order of the expression evaluators so that the generic
expression evaluator is tried first.

rdar://139239935