[lldb] Exclude redundant task flags from summary
Don't show `future` or `childTask` in a Task's summary.

First, `future` is not an official term in Swift Concurrency. There's undetermined value
in showing this. From my reading of the concurrency runtime, in non-embedded Swift all
tasks have the flag `isFuture`. For embedded Swift all tasks do not have the flag set.

Second, child tasks are already indicated with either `groupChildTask`, or
`asyncLetTask`.

Both values are available on demand, ex `p task.isFuture` or `p task.isChildTask`.

This change is to reduce the quantity of data in the summary, because a subsequent
change will include the Tasks name in its summary.