[lldb] Improve determinism of TestSwiftActorUnprioritisedJobs
Change TestSwiftActorUnprioritisedJobs to prevent non-deterministic failures.

1. Removes the sleep, replacing it with an indefinite read from stdin
2. Makes the test not dependent on Foundation
3. Delays the breakpoint by inserting a stalling Task.sleep

The last one needs explanation. The failures happen when the queue (used to serially
execute actor work) has not finished preparing by the time the breakpoint is hit. The
sleep delays the breakpoint from being hit, and that delay should give the queue the
time needed to be in a state needed by this test.