[clang][Frontend] Add overload to ASTPrinter that doesn't own output stream (#142163)
We're planning on using the ASTPrinter in LLDB for AST dumping. But it
currently takes the output stream via `unique_ptr`. In LLDB we don't
have the output stream available in this form and instead it would be
convenient if we could just pass a reference to the stream.

This patch adds that overload.

(cherry picked from commit 41d634397823d01c20db783cf18c9b0c2c41fbf4)