[BoundsSafety] Keep new function type if merged with implicit decl
In BoundsSafety, we allow merging bounds attributes between function
decls if the old decl is implicit. This enables us to add bounds
attributes to a function decl if the old one is implicit without causing
'incompatible function decl' error. For example, we can redeclare the
implicit decl of `memcmp`, and add `__sized_by()` attributes to its
parameters.

In attribute-only mode, those attributes are sugars though, and
compatible function decls are merged by `Context.mergeTypes()`, which
drops those sugars. If both canonical types are the same, we can keep
the new type to retain sugars.

rdar://152192125